+++ tests/lazyProperties/_helpers/_helpers.ts
@@ -1,4 +1,5 @@
-import { LAZY_PROPS, PROXY_KEY } from '../../../src'
+import { LAZY_PROPS } from '../../../src/lazyProperties/trackLazyProps'
+import { PROXY_KEY } from '../../../src/lazyProperties/proxy'
 
 export function getLazyProps(target: Record<PropertyKey, unknown>): undefined | Set<string> {
   return target[LAZY_PROPS] as undefined | Set<string>
